requirements.txt
certifi==2022.12.7
feedparser==6.0.10
logzero==1.7.0
pip==21.3.1
python-dateutil==2.8.2
python-telegram-bot-raw==13.15
setuptools==60.2.0
sgmllib3k==1.0.0
six==1.16.0
wheel==0.37.1
